[qca-nss-dp]: Fix MIB stats truncation in xgmac and qcom HAL
MIB counter struct fields are uint64_t, but the MIB stats population
loop was casting the field pointer to uint32_t *, discarding the upper
32 bits and causing ethtool counters to wrap at 4GB. Fix by using
uint64_t * cast to match the field size.

+++ hal/gmac_ops/qcom/qcom_if.c
@@ -319,7 +319,7 @@ static int32_t qcom_get_eth_stats(struct nss_gmac_hal_dev *nghd, uint64_t *data,
 		p = (uint8_t *)mib_stats
 			+ qcom_gstrings_mib_stats[i_mib].stat_offset;
 		i = QCOM_STATS_LEN + i_mib;
-		data[i] = *(uint32_t *)p;
+		data[i] = *(uint64_t *)p;
 	}
 
 	return 0;
